Enum opentelemetry_otlp::SpanExporter

pub enum SpanExporter {
    Tonic {
        timeout: Duration,
        metadata: Option<MetadataMap>,
        trace_exporter: TonicTraceServiceClient<TonicChannel>,
    },
}

OTLP exporter that sends tracing information

Variants

Tonic

Fields

timeout: Duration

Duration of timeout when sending spans to backend.

metadata: Option<MetadataMap>

Additional headers of the outbound requests.

trace_exporter: TonicTraceServiceClient<TonicChannel>

The Grpc trace exporter

Trace Exporter using tonic as grpc layer.

Implementations

impl SpanExporter

pub fn new_tonic(
config: ExportConfig,
tonic_config: TonicConfig
) -> Result<Self, Error>

Builds a new span exporter with the given configuration.

pub fn from_tonic_channel(
config: ExportConfig,
tonic_config: TonicConfig,
channel: Channel
) -> Result<Self, Error>

Builds a new span exporter with given tonic channel.

This allows users to bring their own custom channel like UDS. However, users MUST make sure the ExportConfig::timeout is the same as the channel’s timeout.