opentelemetry_otlp

Struct TonicExporterBuilder

pub struct TonicExporterBuilder { /* private fields */ }

Configuration for the tonic OTLP GRPC exporter.

It allows you to

add additional metadata

set tls config (via the tls feature)

specify custom channels

Examples

use opentelemetry_sdk::metrics::Temporality;

// Create a span exporter you can use to when configuring tracer providers
let span_exporter = opentelemetry_otlp::SpanExporter::builder().with_tonic().build()?;

// Create a metric exporter you can use when configuring meter providers
let metric_exporter = opentelemetry_otlp::MetricExporter::builder()
    .with_tonic()
    .with_temporality(Temporality::default())
    .build()?;

// Create a log exporter you can use when configuring logger providers
let log_exporter = opentelemetry_otlp::LogExporter::builder().with_tonic().build()?;