Trait opentelemetry_otlp::HasExportConfig

pub trait HasExportConfig {
    fn export_config(&mut self) -> &mut ExportConfig;
}

Provide access to the export config field within the exporter builders.

Required Methods

fn export_config(&mut self) -> &mut ExportConfig

Return a mutable reference to the export config within the exporter builders.